feat: add network status interface to navigation
Add a new navigation item for the network status interface. This commit adds the necessary code changes to include the network status interface in the navigation menu. The network status interface is hidden by default, but can be enabled by setting the environment variable HIDE_NETWORK_STATUS_INTERFACE to false.