refactor: Remove unused code and update function signature in types.ts
Remove the unused import and type definition for MandatoryPluginDetailType in types.ts. Also, update the function signature of handleValidateMandatoryPlugins in PipelineContext to accept optional parameters.

fix: Show mandatory plugin warning in Sidebar

Update the showMandatoryWarning function in Sidebar.tsx to include the check for MandatoryPluginWarning and isJobCard. This ensures that the warning is only shown when necessary.

refactor: Remove unnecessary code in CreatePluginModal.component.tsx

Remove the importComponentFromFELibrary function and the isRequired variable in CreatePluginModal.component.tsx. These are no longer needed.

refactor: Remove unnecessary code in PreBuild.tsx

Remove the importComponentFromFELibrary function and the isRequired variable in PreBuild.tsx. These are no longer needed. Also, update the handlePluginSelection function to use the handleValidateMandatoryPlugins function instead of the validateStage function.